Improve nativeFilters tests
