+++ superset-frontend/cypress-base/cypress/support/index.ts
@@ -283,7 +283,7 @@ Cypress.Commands.add('createSampleCharts', () =>
 
 Cypress.Commands.add(
   'deleteDashboardByName',
-  (dashboardName: string, failOnStatusCode: boolean) =>
+  (dashboardName: string, failOnStatusCode = false) =>
     cy.getDashboards().then((sampleDashboards: any) => {
       const dashboard = sampleDashboards.find(
         (d: any) => d.dashboard_title === dashboardName,
@@ -328,6 +328,38 @@ Cypress.Commands.add('getDashboards', () =>
     .then(resp => resp.body.result),
 );
 
+Cypress.Commands.add('copyDashboard', (originalDashboardName, copyDashboardName) =>
+cy.getDashboards().then((sampleDashboards: any) => {
+  cy.deleteDashboardByName(copyDashboardName, false).then(() => {
+    const dashboard = sampleDashboards.find((d: any) => d.dashboard_title === originalDashboardName);
+    const slugifiedTitle = copyDashboardName.toLowerCase().replace(/ /g, '-');
+    const body = new FormData();
+    body.append("data", JSON.stringify({
+      "dashboard_title": copyDashboardName,
+      "slug": slugifiedTitle,
+      "duplicate_slices": false,
+    }));
+    if (dashboard) {
+        cy
+        .request({
+          method: 'POST',
+          url: `superset/copy_dash/${dashboard.id}/`,
+          body,
+          headers: {
+            'content-type': 'multipart/form-data',
+            Authorization: `Bearer ${TokenName}`,
+          },
+        })
+        .then(() =>
+          cy.getDashboards().then((sampleDashboards: any) => {
+            const newDashboard = sampleDashboards.find((d: any) => d.dashboard_title === copyDashboardName);
+            cy.visit(`/superset/dashboard/${newDashboard.id}/`)
+          })
+        )
+    }
+  })
+}));
+
 Cypress.Commands.add('deleteChart', (id: number, failOnStatusCode = false) =>
   cy
     .request({
@@ -362,7 +394,7 @@ Cypress.Commands.add('getCharts', () =>
 
 Cypress.Commands.add(
   'deleteChartByName',
-  (sliceName: string, failOnStatusCode: boolean) =>
+  (sliceName: string, failOnStatusCode = false) =>
     cy.getCharts().then((sampleCharts: any) => {
       const chart = sampleCharts.find((c: any) => c.slice_name === sliceName);
       if (chart) {
